# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VSM Mobile",
        "description": "Documentation API VSM PK11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "Événements"
                ],
                "summary": "Liste des matchs et entraînements à venir",
                "operationId": "3befe2e8582683a8dc70f7b706a5a97d",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type d'événement (training, match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "training",
                                "match"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendrier récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Event"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Événements"
                ],
                "summary": "Créer un nouvel événement (Admin / Coach)",
                "operationId": "1ed79d9ab6d41b87387cad6237213e5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "type",
                                    "date",
                                    "location"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Match vs FC Akwa"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "match",
                                        "enum": [
                                            "training",
                                            "match"
                                        ]
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-08-30 15:30:00"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Stade PK11, Douala"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Match amical de préparation."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Événement créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Événement créé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "422": {
                        "description": "Données de formulaire invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{id}": {
            "get": {
                "tags": [
                    "Événements"
                ],
                "summary": "Détails d'un événement",
                "operationId": "abf4f4bbfe75221b5bab720a701f17b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Événement introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Événements"
                ],
                "summary": "Modifier un événement (Admin / Coach)",
                "operationId": "ca1f19a988fe9b1ee4af2f4ff3e724fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'événement à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Entraînement déplacé"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "training",
                                        "enum": [
                                            "training",
                                            "match"
                                        ]
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-08-22 07:00:00"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Stade PK11, Douala"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Changement d'horaire exceptionnel."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Événement mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Événement mis à jour avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Événement introuvable"
                    },
                    "422": {
                        "description": "Données de formulaire invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Événements"
                ],
                "summary": "Supprimer un événement (Admin / Coach)",
                "operationId": "54be98df9381a8b89013d29e170a31fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'événement à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Événement supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Événement supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Événement introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{id}/presence": {
            "post": {
                "tags": [
                    "Événements"
                ],
                "summary": "Confirmer ou décliner sa présence à un événement",
                "operationId": "9aeac98bc81671ab109580e0830998fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'événement (Match / Entraînement)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Statut de présence du joueur",
                                        "type": "string",
                                        "example": "present",
                                        "enum": [
                                            "present",
                                            "absent",
                                            "uncertain"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut de présence mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Votre présence a été enregistrée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Événement introuvable"
                    },
                    "422": {
                        "description": "Statut de présence invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Event": {
                "title": "Événement",
                "description": "Modèle représentant un événement (match ou entraînement)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Entraînement hebdomadaire"
                    },
                    "type": {
                        "type": "string",
                        "example": "training",
                        "enum": [
                            "training",
                            "match"
                        ]
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-22 06:00:00"
                    },
                    "location": {
                        "type": "string",
                        "example": "Stade PK11, Douala"
                    },
                    "description": {
                        "type": "string",
                        "example": "Séance d'entraînement physique et mise en place tactique.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Entrez votre jeton Sanctum sous la forme: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}